investigation search consolidation
