+++ apps/modernization-api/src/main/resources/graphql/shared.graphqls
@@ -13,38 +13,6 @@ input SortablePage {
   sortDirection: SortDirection
 }
 
-type PersonParticipation {
-  actUid: Int!
-  localId: String
-  typeCd: String
-  entityId: Int!
-  subjectClassCd: String
-  participationRecordStatus: String
-  typeDescTxt: String
-  participationLastChangeTime: DateTime
-  firstName: String
-  lastName: String
-  birthTime: DateTime
-  currSexCd: String
-  personCd: String!
-  personParentUid: Int
-  personRecordStatus: String!
-  personLastChangeTime: DateTime
-  shortId: Int
-}
-
-type OrganizationParticipation {
-  actUid: Int
-  typeCd: String
-  entityId: Int
-  subjectClassCd: String
-  typeDescTxt: String
-  participationRecordStatus: String
-  participationLastChangeTime: DateTime
-  name: String
-  organizationLastChangeTime: DateTime
-}
-
 enum SortField {
   lastNm
   birthTime
